Why:
- PR #544 had diverged from main's resource-policy and decode hardening work.

What:
- merge origin/main while retaining Dragline compiler, profiling, tiering, and cache identity semantics
- enforce main's module, native-code, instance-metadata, and memory resource policies
- regenerate the public facade and preserve the combined native ABI accounting

Proof:
- make test
- go test ./cli/runtime/internal/artifactcache ./src/core/compiler/backend/railshot/shared ./src/wago
- go test ./src/core/runtime ./src/wago

Docs:
- merged main's runtime resource-model and limit documentation

+++ src/wago/gc_struct_globals_amd64_test.go
@@ -54,7 +54,7 @@ func TestStagedGCStructGlobalRootsAndPublicEgress(t *testing.T) {
 			if state == nil || state.gcGlobalRootCount != 2 {
 				t.Fatalf("GC global root mapping = %#v", state)
 			}
-			for i := uint8(0); i < state.gcGlobalRootCount; i++ {
+			for i := uint32(0); i < state.gcGlobalRootCount; i++ {
 				mapping := state.gcGlobalRoots[i]
 				want := gc.Ref(uint32(readGlobalObject(in.globalCells[mapping.GlobalIndex], ValAnyRef)))
 				got, err := collector.CheckedGlobalSlot(mapping.SlotIndex)
